Revert "feat：add slider (#1178)"
This reverts commit 7eb83edd52b561c1af02d6e799051716453636b2.
