+++ src/mip-story/mip-story-view.js
@@ -40,8 +40,8 @@ define(function (require) {
 
     customElement.prototype.toggleAllMedia = function (e, muted) {
         this.muted = muted;
-        var ele = e.target || null;
-        if (ele && ele.hasAttribute && ele.hasAttribute('muted')) {
+        var ele = e.target;
+        if (ele.hasAttribute('muted')) {
             !this.muted && this.resumeAllMedia();
             !this.muted && this.unMuteAllMedia();
         }
@@ -68,106 +68,23 @@ define(function (require) {
         });
     };
 
-    customElement.prototype.setPreActive = function (eventEmiter) {
-        this.parentEmiter = eventEmiter;
-        this.animationElements = [];
-        this.initAnimationFirstFrame();
-        // css-获取有动画的节点，并且放到数组中便于修改display
-        this.findAnimationNodes(this.element);
-        // css-修改每个有动画节点的display
-        this.initFirstFrameStyle(false);
-    };
 
-    // 监控CSS中是否有动画
-    function hasCssAnimation(obj) {
-        var ani = null;
-        try {
-            ani = document.defaultView.getComputedStyle(obj)['animationName']
-                || document.defaultView.getComputedStyle(obj)['-webkit-animationName'];
-        } catch (e) {
-        }
-        if (ani && ani != 'none') {
-            return true;
-        }
-        return false;
-    }
-
-    customElement.prototype.findAnimationNodes = function (parent) {
-        if (parent == null) return;
-        var subNodes = parent.children;
-        for (var index = 0; index < subNodes.length; index++) {
-            if (hasCssAnimation(subNodes[index])) {
-                this.animationElements.push(subNodes[index]);
-            }
-            if (subNodes[index].children.length > 0) {
-                this.findAnimationNodes(subNodes[index]);
-            }
-        }
-    };
-
-    function toggleDisplay(obj, disp) {
-        if (disp) {
-            obj.setAttribute('style', 'display: ' + obj.getAttribute("originDisplay"));
-        } else {
-            var originDisplay = document.defaultView.getComputedStyle(obj)['display'];
-            obj.setAttribute('originDisplay', originDisplay);
-            obj.setAttribute('style', 'display: none');
-        }
-    }
-
-    customElement.prototype.initFirstFrameStyle = function (disp) {
-        if (this.animationElements != null) {
-            for (var index = 0; index < this.animationElements.length; index++) {
-                toggleDisplay(this.animationElements[index], disp);
-            }
-        }
-    };
-
-    customElement.prototype.setAllMedia = function (status, muted, load, eventEmiter, viewType) {
+    customElement.prototype.setActive = function (status, muted, load, eventEmiter) {
         this.muted = muted;
         this.parentEmiter = eventEmiter;
         if (status) {
-            this.initAnimationFirstFrame();
+            this.element.setAttribute('active', '');
             this.maybeStartAnimation();
             this.resumeAllMedia(load);
             this.muted ? this.muteAllMedia() : this.unMuteAllMedia();
         }
         else {
+            this.element.removeAttribute('active');
             this.maybeClearAutoAdvance();
             this.pauseAllMedia();
             this.maybeClearAnimation();
         }
     };
-    customElement.prototype.setCssMedia = function (status, muted, eventEmiter) {
-        this.muted = muted;
-        this.parentEmiter = eventEmiter;
-        if (status) {
-            this.initFirstFrameStyle(true);
-        }
-        else {
-            this.initFirstFrameStyle(false);
-        }
-    };
-
-    customElement.prototype.clearCssMedia = function (status, muted, eventEmiter) {
-        if (this.animationElements != null) {
-            for (var index = 0; index < this.animationElements.length; index++) {
-                toggleDisplay(this.animationElements[index], true);
-                this.animationElements[index].removeAttribute('originDisplay')
-            }
-        }
-    };
-    customElement.prototype.initAnimationFirstFrame = function () {
-        if (hasAnimations(this.element)) {
-            css(this.element, {visibility: 'hidden'});
-            if (!this.animationManager) {
-                this.animationManager = new AnimationManager(this.element);
-            }
-            this.animationManager.paintFirstFrame();
-            css(this.element, {visibility: ''});
-        }
-    };
-
     customElement.prototype.maybeStartAnimation = function () {
         if (hasAnimations(this.element)) {
             css(this.element, {visibility: 'hidden'});
@@ -179,10 +96,9 @@ define(function (require) {
             this.animationManager.runAllAnimate();
             this.maybeSetAutoAdvance();
         }
-
     };
 
-    customElement.prototype.maybeClearAnimation = function () {
+    customElement.prototype.maybeClearAnimation = function() {
         if (this.animationManager) {
             this.animationManager.cancelAllAnimate();
         }
@@ -209,7 +125,6 @@ define(function (require) {
     customElement.prototype.initView = function () {
         this.audio = new Audio();
         var node = this.element.parentNode;
-        this.animationElements = [];
         if (!node.hasAttribute(BACKGROUND_AUDIO)) {
             var audioSrc = this.element.getAttribute(BACKGROUND_AUDIO);
             this.audio.build(this.element, audioSrc);
